requirements.txt
h5py==3.12.1
tensorflow==2.17.0
matplotlib==3.9.2
pyyaml==6.0.2
dvc==3.53.2
numpy~=1.26.4
scikit-learn~=1.5.2
seaborn~=0.13.2
bentoml~=1.3.14
pillow~=11.0.0
pydantic~=2.9.2